Cleaned up warnings
Procrastinated work to make my output seem nicer :^)

+++ DX11 Framework 2018 Uni/GameTimer.cpp
@@ -12,7 +12,7 @@ GameTimer::GameTimer()
 	previousTime = 0;
 	currentTime = 0;
 	timerStopped = false;
-	stopTime = 0.0f;
+	stopTime = (__int64)0.0f;
 
 	__int64 countsPerSec;
 	QueryPerformanceFrequency((LARGE_INTEGER*)&countsPerSec);
